Update no-std example.
