+++ .github/workflows/ci.yml
@@ -9,7 +9,7 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         tc: [stable, beta, nightly]
-        ar: [--all, --all --features=no-std]
+        ar: [--all, --all --no-default-features]
     steps:
     - uses: actions/checkout@v2
     - uses: actions-rs/toolchain@v1
@@ -47,7 +47,7 @@ jobs:
     - uses: actions-rs/cargo@v1
       with:
         command: build
-        args: --features=no-std --target=${{ matrix.cc }}
+        args: --no-default-features --target=${{ matrix.cc }}
   cross-compile-ios:
     runs-on: ${{ matrix.os }}
     strategy:
@@ -66,7 +66,7 @@ jobs:
     - uses: actions-rs/cargo@v1
       with:
         command: build
-        args: --features=no-std --target=${{ matrix.cc }}
+        args: --no-default-features --target=${{ matrix.cc }}
   cross-compile-web:
     runs-on: ${{ matrix.os }}
     strategy:
